requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
arrow==1.2.2
async-timeout==4.0.2
attrs==21.4.0
boto3==1.20.52
botocore==1.23.52
bravado==11.0.3
bravado-core==5.17.0
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
comet-ml==3.27.0
configobj==5.0.6
cycler==0.11.0
datasets==1.18.3
dill==0.3.4
dulwich==0.20.32
everett==3.0.0
fever-drqa==1.0.13
filelock==3.4.2
fonttools==4.29.1
fqdn==1.5.1
frozenlist==1.3.0
fsspec==2022.1.0
future==0.18.2
gitdb==4.0.9
GitPython==3.1.26
huggingface-hub==0.4.0
idna==3.3
importlib-resources==5.4.0
isoduration==20.11.0
jmespath==0.10.0
joblib==1.1.0
jsonpointer==2.2
jsonref==0.2
jsonschema==4.4.0
kiwisolver==1.3.2
matplotlib==3.5.1
monotonic==1.6
msgpack==1.0.3
multidict==6.0.2
multiprocess==0.70.12.2
neptune-client==0.14.3
nltk==3.6.7
numpy==1.22.2
nvidia-ml-py3==7.352.0
oauthlib==3.2.0
packaging==21.3
pandas==1.4.0
pexpect==4.8.0
Pillow==9.0.1
prettytable==3.0.0
psutil==5.9.0
ptyprocess==0.7.0
pyarrow==7.0.0
PyJWT==2.3.0
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
regex==2022.1.18
requests==2.27.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rfc3339-validator==0.1.4
rfc3987==1.3.8
s3transfer==0.5.1
sacremoses==0.0.47
scikit-learn==1.0.2
scipy==1.8.0
semantic-version==2.9.0
simplejson==3.17.6
six==1.16.0
smmap==5.0.0
swagger-spec-validator==2.7.4
termcolor==1.1.0
threadpoolctl==3.1.0
tokenizers==0.11.4
torch==1.10.2
tqdm==4.62.3
transformers==4.16.2
typing_extensions==4.0.1
uri-template==1.1.0
urllib3==1.26.8
wcwidth==0.2.5
webcolors==1.11.1
websocket-client==1.2.3
wrapt==1.13.3
wurlitzer==3.0.2
xxhash==2.0.2
yarl==1.7.2
zipp==3.7.0